CB-20411 Remove more circular dependency
2 cases are handled:
```
lowLogsToListDiagnosticsCollectionResponseConverter (field private com.sequenceiq.flow.core.config.FlowProgressHolder com.sequenceiq.cloudbreak.converter.v4.diagnostics.FlowLogsToListDiagnosticsCollectionResponseConverter.flowProgressHolder)
┌─────┐
flowProgressHolder defined in URL [jar:file:/cloudbreak.jar!/BOOT-INF/lib/flow-2.67.0-b102.jar!/com/sequenceiq/flow/core/config/FlowProgressHolder.class]
↑ ↓
rotateSaltPasswordFlowConfig
↑ ↓
rotateSaltPasswordActions (field private com.sequenceiq.cloudbreak.service.RotateSaltPasswordService com.sequenceiq.cloudbreak.core.flow2.cluster.salt.rotatepassword.RotateSaltPasswordActions.rotateSaltPasswordService)
↑ ↓
rotateSaltPasswordService (field private com.sequenceiq.cloudbreak.core.flow2.service.ReactorFlowManager com.sequenceiq.cloudbreak.service.RotateSaltPasswordService.flowManager)
↑ ↓
reactorFlowManager (field private com.sequenceiq.cloudbreak.core.flow2.service.TerminationTriggerService com.sequenceiq.cloudbreak.core.flow2.service.ReactorFlowManager.terminationTriggerService)
↑ ↓
terminationTriggerService (field private com.sequenceiq.flow.service.FlowCancelService com.sequenceiq.cloudbreak.core.flow2.service.TerminationTriggerService.flowCancelService)
↑ ↓
flowCancelService (field private com.sequenceiq.flow.core.Flow2Handler com.sequenceiq.flow.service.FlowCancelService.flow2Handler)
↑ ↓
flow2Handler
↑ ↓
clusterCreationFlowConfig
↑ ↓
clusterCreationActions (field private com.sequenceiq.cloudbreak.job.stackpatcher.ExistingStackPatcherJobService com.sequenceiq.cloudbreak.core.flow2.cluster.provision.ClusterCreationActions.existingStackPatcherJobService)
↑ ↓
existingStackPatcherJobService (field private com.sequenceiq.cloudbreak.job.stackpatcher.ExistingStackPatcherServiceProvider com.sequenceiq.cloudbreak.job.stackpatcher.ExistingStackPatcherJobService.existingStackPatcherServiceProvider)
↑ ↓
existingStackPatcherServiceProvider (field private java.util.Collection com.sequenceiq.cloudbreak.job.stackpatcher.ExistingStackPatcherServiceProvider.existingStackPatchServices)
↑ ↓
unboundRestartPatchService (field private com.sequenceiq.flow.service.FlowService com.sequenceiq.cloudbreak.service.stackpatch.UnboundRestartPatchService.flowService)
↑ ↓
flowService (field private com.sequenceiq.flow.converter.FlowProgressResponseConverter com.sequenceiq.flow.service.FlowService.flowProgressResponseConverter)
↑ ↓
flowProgressResponseConverter defined in URL [jar:file:/cloudbreak.jar!/BOOT-INF/lib/flow-2.67.0-b102.jar!/com/sequenceiq/flow/converter/FlowProgressResponseConverter.class]
└─────┘
```
and
```
{"timestamp":"2023-01-26 14:18:03.443","level":"ERROR","thread":"main","logger":"org.springframework.boot.diagnostics.LoggingFailureAnalysisReporter","message":"2023-01-26 14:18:03,443 [main] report:40 ERROR o.s.b.d.LoggingFailureAnalysisReporter - [type:springLog] [crn:] [name:] [flow:] [requestid:] [tenant:] [userCrn:] [environment:] [traceId:] [spanId:] \n\n***************************\nAPPLICATION FAILED TO START\n***************************\n\nDescription:\n\nThe dependencies of some of the beans in the application context form a cycle:\n\n   progressController (field private com.sequenceiq.datalake.service.sdx.progress.ProgressService com.sequenceiq.datalake.controller.progress.ProgressController.progressService)\n      ↓\n   progressService (field private com.sequenceiq.flow.service.FlowService com.sequenceiq.datalake.service.sdx.progress.ProgressService.flowService)\n      ↓\n   flowService\n┌─────┐\n|  failHandledEvents defined in class path resource [com/sequenceiq/flow/core/config/Flow2Config.class]\n↑     ↓\n|  updateLoadBalancerDNSFlowConfig\n↑     ↓\n|  updateLoadBalancerDNSActions (field private com.sequenceiq.datalake.service.sdx.SdxService com.sequenceiq.datalake.flow.loadbalancer.dns.UpdateLoadBalancerDNSActions.sdxService)\n↑     ↓\n|  sdxService (field private com.sequenceiq.flow.service.FlowCancelService com.sequenceiq.datalake.service.sdx.SdxService.flowCancelService)\n↑     ↓\n|  flowCancelService (field private com.sequenceiq.flow.core.Flow2Handler com.sequenceiq.flow.service.FlowCancelService.flow2Handler)\n↑     ↓\n|  flow2Handler\n└─────┘\n\n\n"}
```
